requirements.txt
GitPython==3.1.42
Pillow==10.4.0
Requests==2.32.3
requests_toolbelt==1.0.0
streamlit==1.32.2
tqdm==4.66.4
websocket_client==1.7.0
